support fractional share quantities; fix cash save for lowercase currency
  Change Stock.amount from Int to Double throughout the stack so users can
  hold fractional shares (e.g. 100.5). Whole-number quantities continue to
  display without a trailing ".0" (1163 not 1163.0).

  Also fix cash.txt save failing to match entries whose currency is written
  in lowercase (e.g. "Loan.hkd=123"): normalize the file key before looking
  it up in the update map via normalizeCashKey().

  - model/Stock.kt: amount Int → Double
  - service/CsvStockReader.kt: parse amount with toDouble()
  - web/PortfolioRenderer.kt: formatQty helper; step="any" on qty input
  - web/routes.kt: formatQty helper; Pair<Int,Double> → Pair<Double,Double>;
    parse JSON amount as double; write clean qty strings to CSV;
    normalizeCashKey() on cash.txt key lookup
  - static/portfolio.js: parseInt → parseFloat for all qty reads;
    remove Math.round from Google Sheets paste handler

+++ src/main/kotlin/com/portfoliohelper/service/CsvStockReader.kt
@@ -30,7 +30,7 @@ object CsvStockReader {
             for (record in parser) {
                 try {
                     val label = record.get("stock_label")
-                    val amount = record.get("amount").toInt()
+                    val amount = record.get("amount").toDouble()
 
                     // Read target_weight if column exists (backward compatible)
                     val targetWeight = try {
